pyproject.toml
[project]
name = "patient-deduplication"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "fuzzy",
    "phonenumbers"
]
